requirements.txt
alembic==1.4.2
appdirs==1.4.3
atomicwrites==1.4.0
attrs==19.3.0
cachelib==0.1
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.1
colorama==0.4.3
cssselect==1.1.0
Flask==1.1.1
Flask-Migrate==2.5.3
Flask-SQLAlchemy==2.4.1
Flask-WTF==0.14.3
howdoi==1.2.1
idna==2.9
importlib-metadata==1.6.0
itsdangerous==1.1.0
Jinja2==2.11.1
lxml==4.5.0
Mako==1.1.2
MarkupSafe==1.1.1
more-itertools==8.3.0
packaging==20.4
pluggy==0.13.1
py==1.8.1
Pygments==2.6.1
pyparsing==2.4.7
pyquery==1.4.1
pytest==5.4.2
python-dateutil==2.8.1
python-dotenv==0.13.0
python-editor==1.0.4
requests==2.23.0
six==1.14.0
SQLAlchemy==1.3.16
urllib3==1.25.9
wcwidth==0.1.9
Werkzeug==1.0.1
WTForms==2.3.1
zipp==3.1.0
